+++ miniapp/src/main/java/com/rakuten/tech/mobile/miniapp/AdMob.kt
@@ -1,5 +1,6 @@
-package com.rakuten.tech.mobile.miniapp.ads
+package com.rakuten.tech.mobile.miniapp
 
+import android.util.Log
 import androidx.annotation.VisibleForTesting
 
 /**
@@ -14,6 +15,17 @@ internal fun isAdMobProvided(): Boolean =
     } catch (e: ClassNotFoundException) {
         false
     }
+
+/** Check whether hostapp provides admob dependency. */
+@Suppress("EmptyCatchBlock", "SwallowedException")
+internal inline fun <T> whenHasH5AdsWebViewClient(callback: () -> T) {
+    try {
+        Class.forName("com.rakuten.tech.mobile.miniapp.ads.MiniAppH5AdsProvider")
+        callback.invoke()
+    } catch (e: ClassNotFoundException) {
+        Log.e("Missing Dependency", ":admob-latest")
+    }
+}
 
 @VisibleForTesting
 internal var AdMobClassName = "com.google.android.gms.ads.MobileAds"