+++ testapp/src/main/java/com/rakuten/tech/mobile/testapp/BuildVariant.kt
@@ -6,4 +6,3 @@ enum class BuildVariant(val value: String){
     RC("rc"),
     RELEASE("release")
 }
-