+++ testapp/src/main/java/com/rakuten/tech/mobile/testapp/ui/display/MiniAppDisplayViewModelFactory.kt
@@ -5,8 +5,9 @@ import androidx.lifecycle.ViewModelProvider
 import com.rakuten.tech.mobile.miniapp.MiniApp
 
 @Suppress("UNCHECKED_CAST")
-class MiniAppDisplayViewModelFactory(private val miniApp: MiniApp): ViewModelProvider.NewInstanceFactory(){
-    override fun <T : ViewModel?> create(modelClass: Class<T>): T {
+class MiniAppDisplayViewModelFactory(private val miniApp: MiniApp) :
+    ViewModelProvider.NewInstanceFactory() {
+    override fun <T : ViewModel> create(modelClass: Class<T>): T {
         return MiniAppDisplayViewModel(miniapp = miniApp) as T
     }
 }